requirements.txt
Bottleneck==1.4.2
Brotli==1.0.9
PySocks==1.7.1
PyYAML==6.0.2
SQLAlchemy==2.0.34
appnope==0.1.4
asttokens==2.4.1
boto3==1.34.154
botocore==1.34.154
certifi==2024.8.30
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
greenlet==3.0.1
idna==3.7
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jmespath==1.0.1
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numexpr==2.10.1
numpy==1.26.4
packaging==24.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.2
platformdirs==4.3.6
prompt-toolkit==3.0.48
psutil==5.9.0
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
requests==2.32.3
s3transfer==0.10.1
setuptools==75.1.0
six==1.16.0
stack-data==0.6.2
tabula-py==2.10.0
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.11.0
tzdata==2023.3
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
zipp==3.21.0
autocommand==2.2.2
backports.tarfile==1.2.0
importlib-resources==6.4.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
tomli==2.0.1
typeguard==4.3.0
